renaming cache modification utils file
