+++ web/src/cache-queries/current-user-posts-cursor.ts
@@ -1,4 +1,4 @@
-import { gql } from "@apollo/client";
+import { gql } from '@apollo/client';
 
 const getCurrentGetUserPostsCursor = gql`
   query GetCurrentGetUserPostsCursor {
