+++ web/src/cache-queries/current-user-profile.ts
@@ -1,4 +1,4 @@
-import { gql } from "@apollo/client";
+import { gql } from '@apollo/client';
 
 const getCurrentUserProfile = gql`
   query GetCurrentUserProfile {
