+++ web/src/components/comment-input/index.tsx
@@ -1,10 +1,9 @@
-import React, { ChangeEvent } from 'react';
+import React, { ChangeEvent, useState } from 'react';
 import { TextField } from '@material-ui/core';
 import { useMutation } from '@apollo/client';
 import { CreateCommentDocument } from '../../generated/graphql';
-import { useState } from 'react';
 import { currentUserProfileVar } from '../../cache';
-import { readGetUserPostsQuery, updatePosts } from '../../utils/update-comments';
+import { readGetUserPostsQuery, updatePosts } from '../../utils/cache-modification.util';
 
 interface Props {
   userId: number,
